+++ src/frontend/apps/desk/src/features/mail-domains/access-management/api/index.ts
@@ -1,5 +1,7 @@
 export * from './useMailDomainAccesses';
+export * from './useInvitationMailDomainAccesses';
 export * from './useUpdateMailDomainAccess';
 export * from './useCreateMailDomainAccess';
 export * from './useDeleteMailDomainAccess';
+export * from './useDeleteMailDomainInvitation';
 export * from './useCreateInvitation';